backend/requirements.txt
fastapi
uvicorn
sqlmodel
python-multipart
python-jose[cryptography]
passlib[bcrypt]
bcrypt<5.0.0
psycopg2-binary
python-dotenv
pydantic-settings
openai>=1.0.0
